VocalRAG/asr-service/requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
python-multipart==0.0.20
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
numpy<2.0.0